requirements.txt
# Config
python-dotenv>=1.0.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Browser automation for LinkedIn
playwright>=1.41.0

# Web UI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
